close #8 & close #6
